backend/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
confluent-kafka==1.7.0
idna==3.3
kafka-python==2.0.2
requests==2.26.0
urllib3==1.26.7
Flask==2.0.1
Flask-Cors==3.0.10
requests-unixsocket==0.2.0
responses==0.5.0
pandas>=1.2.4
statistics>=1.0.3.5